requirements.txt
amadeus==12.0.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
deprecation==2.1.0
fastapi==0.116.1
Flask==3.1.1
flask-cors==6.0.1
flask-swagger-ui==5.21.0
flatbuffers==25.2.10
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==2.12.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
jh2==5.0.9
Jinja2==3.1.6
MarkupSafe==3.0.2
niquests==3.14.1
numpy==2.3.2
openmeteo_requests==1.6.0
openmeteo_sdk==1.20.1
packaging==25.0
platformdirs==4.3.8
postgrest==1.1.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
qh3==1.5.3
realtime==2.6.0
requests==2.32.4
requests-cache==1.2.1
retry-requests==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.5.0
urllib3-future==2.13.900
uvicorn==0.35.0
wassima==1.2.2
websockets==15.0.1
Werkzeug==3.1.3